Show pattern summary on shared page too
The aggregated pattern card ("9 unique · 30 total" with per-pattern
count + quality) was only rendered on the authenticated analyze
page. The public /shared route was missing it entirely, so
recipients of a share link saw the score + timeline + strengths/
improvements but no per-pattern breakdown.

Extracted PatternSummaryCard and derivePatternSummary from
app/(app)/analyze/page.tsx into src/components/analyze/pattern-
summary.tsx and imported both on the analyze page and the shared
page. Shared page applies the same "server summary if present,
else derive client-side" fallback so historical analyses still
render the card.

No visual change on the analyze page; the shared page gains a
Patterns card between the timeline and Strengths.
